backend/requirements.txt
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.10.1
mpmath==1.1.0
numpy==1.16.4
pytz==2019.1
sqlparse==0.3.0
sympy==1.4
